Move some math tests to lualib
